requirements.txt
# torch
# torchvision
numpy
Pillow
onnx
onnxruntime
tqdm
flask~=1.1.2